requirements.txt
fastapi>=0.79.0
aiogram>=2.21
python-dotenv>=0.20.0
